+++ pkg/kepval/keps/proposals.go
@@ -26,63 +26,14 @@ import (
 
 	"github.com/pkg/errors"
 	"gopkg.in/yaml.v2"
+
+	"k8s.io/enhancements/api"
 	"k8s.io/enhancements/pkg/kepval/keps/validations"
 )
 
-type Proposals []*Proposal
-
-func (p *Proposals) AddProposal(proposal *Proposal) {
-	*p = append(*p, proposal)
-}
-
-type Milestone struct {
-	Alpha  string `json:"alpha" yaml:"alpha"`
-	Beta   string `json:"beta" yaml:"beta"`
-	Stable string `json:"stable" yaml:"stable"`
-}
-
-type FeatureGate struct {
-	Name       string   `json:"name" yaml:"name"`
-	Components []string `json:"components" yaml:"components"`
-}
-
-type Proposal struct {
-	ID       string `json:"id"`
-	PRNumber string `json:"prNumber,omitempty"`
-	Name     string `json:"name,omitempty"`
-
-	Title             string   `json:"title" yaml:"title"`
-	Number            string   `json:"kep-number" yaml:"kep-number"`
-	Authors           []string `json:"authors" yaml:",flow"`
-	OwningSIG         string   `json:"owningSig" yaml:"owning-sig"`
-	ParticipatingSIGs []string `json:"participatingSigs" yaml:"participating-sigs,flow,omitempty"`
-	Reviewers         []string `json:"reviewers" yaml:",flow"`
-	Approvers         []string `json:"approvers" yaml:",flow"`
-	PRRApprovers      []string `json:"prrApprovers" yaml:"prr-approvers,flow"`
-	Editor            string   `json:"editor" yaml:"editor,omitempty"`
-	CreationDate      string   `json:"creationDate" yaml:"creation-date"`
-	LastUpdated       string   `json:"lastUpdated" yaml:"last-updated"`
-	Status            string   `json:"status" yaml:"status"`
-	SeeAlso           []string `json:"seeAlso" yaml:"see-also,omitempty"`
-	Replaces          []string `json:"replaces" yaml:"replaces,omitempty"`
-	SupersededBy      []string `json:"supersededBy" yaml:"superseded-by,omitempty"`
-
-	Stage           string    `json:"stage" yaml:"stage"`
-	LatestMilestone string    `json:"latestMilestone" yaml:"latest-milestone"`
-	Milestone       Milestone `json:"milestone" yaml:"milestone"`
-
-	FeatureGates     []FeatureGate `json:"featureGates" yaml:"feature-gates"`
-	DisableSupported bool          `json:"disableSupported" yaml:"disable-supported"`
-	Metrics          []string      `json:"metrics" yaml:"metrics"`
-
-	Filename string `json:"-" yaml:"-"`
-	Error    error  `json:"-" yaml:"-"`
-	Contents string `json:"markdown" yaml:"-"`
-}
-
 type Parser struct{}
 
-func (p *Parser) Parse(in io.Reader) *Proposal {
+func (p *Parser) Parse(in io.Reader) *api.Proposal {
 	scanner := bufio.NewScanner(in)
 	count := 0
 	metadata := []byte{}
@@ -99,7 +50,7 @@ func (p *Parser) Parse(in io.Reader) *Proposal {
 			body.WriteString(line)
 		}
 	}
-	proposal := &Proposal{
+	proposal := &api.Proposal{
 		Contents: body.String(),
 	}
 	if err := scanner.Err(); err != nil {
